+++ src/db.ts
@@ -1,4 +1,4 @@
-import { readFileSync, existsSync } from 'fs';
+import { readFileSync } from 'node:fs';
 import {
   BenchmarkData,
   Executor as ApiExecutor,
